2012.10.20 Portolás 64 bitre Windows 2012-n (második lépés). Első lépésben elértük, hogy minden forduljon és fusson, ami 32 biten futott. A második lépés célja, hogy a programok ki tudják használni a nagy címteret, azaz működjön a navigáció a filemapping nagy fájlokra is. (Unixon ezek korábban is működtek, csak Windowson van működésbeli változás.) Binary stringek hossza unsigned long -> binarysize_t (=size_t). Új makrók:  D2LONGW(x), D2ULONGW(x), D2ULONGX(x). Új makrók: PARNLW(i), PARNUW(i), PARNUX(i). Új makrók: _parnlw(i), _parnuw(i), _parnux(i). trim,  substr, slice, empty, at, rat, array: átírva binarysize_t-re. filemmap.cpp: átírva binarysize_t-re hibakezelés javítva. ftruncate: új modul fájl méret állítására (Linux/Windows külön). chsize: visszavezetve ftruncate-ra (depreacted, compatibility). fsetlock:  Windows/Linux szétválasztva, Windowson újraírva. fseek: Windowson újraítva _lseeki64()-re. findnext: fájl dátum kiolvasás újraírva, méret kilvasás javítva. vmm.cpp: átírva binarysize_t-re.

+++ ccclib/src/varprint.cpp
@@ -97,7 +97,7 @@ void var_print(VALUE *v)
         case TYPE_BINARY:
         {
             OUTSTR(L"BINARY length=");
-            OUTNUM(BINARYLEN(v));
+            OUTNUM((double)BINARYLEN(v));
             OUTOREF(v->data.binary.oref);
             
             if(BINARYPTR0(v)==0)
